moved password hash to new file
