+++ project/src/main/java/com/google/sps/classes/PasswordHash.java
@@ -1,0 +1,29 @@
+package com.google.sps.classes;
+import org.apache.commons.codec.binary.Hex;
+import javax.crypto.SecretKey;
+import javax.crypto.SecretKeyFactory;
+import javax.crypto.spec.PBEKeySpec;
+import java.io.UnsupportedEncodingException;
+import java.security.NoSuchAlgorithmException;
+import java.security.spec.InvalidKeySpecException;
+
+
+
+public class PasswordHash{
+    public static String hashPassword(final char[] password){
+        try {
+                String salt = "@*1!";
+                int iterations = 500;
+                int keyLength = 412;
+                byte[] saltBytes = salt.getBytes();
+                SecretKeyFactory skf = SecretKeyFactory.getInstance( "PBKDF2WithHmacSHA512" );
+                PBEKeySpec spec = new PBEKeySpec( password, saltBytes, iterations, keyLength );
+                SecretKey key = skf.generateSecret( spec );
+                byte[] res = key.getEncoded( );
+                return Hex.encodeHexString(res);
+            } catch ( NoSuchAlgorithmException | InvalidKeySpecException e ) {
+                throw new RuntimeException( e );
+            }
+    }
+
+}
