+++ project/src/main/webapp/login.html
@@ -5,7 +5,7 @@
         <title>Log In</title>
         <link rel="stylesheet" href="icons.css">
         <link rel="stylesheet" href="materialize.min.css">
-       <script src="materialize.min.js"></script>
+        <script src="materialize.min.js"></script>
         <script type="text/javascript" src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
         <script>
         $(document).ready(function(){
@@ -14,6 +14,7 @@
         </script>
     </head>
     <body onload = "">
+        <script type="text/javascript" src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
         <nav>
     	<div class="nav-wrapper blue-grey lighten-4">
             <a href="#" data-target="mobile-demo" class="sidenav-trigger"><i style="color: #42a5f5;" class="material-icons">menu</i></a>
@@ -59,11 +60,12 @@
                 <div class="row">
                     <div class="col s12 l8 offset-l2">
                         <div class="input-field right">
-                            <button class="btn white-text" style="background-color: #42a5f5;">Log In</button>
+                            <button onlick="check()" class="btn white-text" style="background-color: #42a5f5;">Log In</button>
                         </div>
                     </div>
                 </div>
             </form>
+            <div id="error"></div>
             <div class="row">
                 <div class="col s12 l8 offset-l2">
                     <h6>Don't have an account? Create an account <a href="/signup.html">here</a></h6>
