+++ project/src/main/webapp/signup.html
@@ -86,7 +86,7 @@
                 </div>
                 <div class="row">
                     <div class="input-field col s12">
-                        <input type="password" name="password" class="validate" id="password" required>
+                        <input type="password" name="password" class="validate" id="password" pattern=".{8,}" title="Eight or more characters"required>
                         <label for="password">Password</label>
                     </div>
                 </div>
